requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
asyncpg==0.29.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.0
langchain==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langgraph==0.2.72
langsmith==0.1.129
groq==0.9.0
sentence-transformers==2.7.0
transformers==4.41.2
protobuf==4.25.3
faiss-cpu==1.8.0
pypdf==4.2.0
twilio==9.0.5
aiosqlite==0.20.0